Refactored RecordScriptureFragment and RecordResourceFragment to not extend RecordableFragment
